+++ spec/rubocop/cop/salsify/rails_application_serializer_spec.rb
@@ -1,7 +1,7 @@
 # encoding utf-8
 # frozen_string_literal: true
 
-describe RuboCop::Cop::Salsify::DandelionApplicationSerializer, :config do
+describe RuboCop::Cop::Salsify::RailsApplicationSerializer, :config do
   let(:msgs) { [described_class::MSG] }
   let(:highlights) { ['ActiveModel::Serializer'] }